pyproject.toml
[tool.poetry]
name = "use-cases"
version = "0.1.0"
description = ""
authors = ["PibZuncy <noma.kazuhiro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.28.1"
langchain = "^0.0.315"
playwright = "^1.38.0"
beautifulsoup4 = "^4.12.2"
ipykernel = "^6.26.0"
html2text = "^2020.1.16"
notebook = "^7.0.6"
python-dotenv = "1.0.0"
tiktoken = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
